refactor: absorb animations into presentation package (5.17.1b)
- Move animations source → presentation/src/animations/
- Add ./animations subpath export to presentation
- Update docs app to import from @revealui/presentation/animations
- Remove @revealui/animations package (only consumed by docs showcase)
- All 917 presentation tests pass (including 8 migrated animation tests)

Problem solved: animations serve UI components, not independently useful.
Presentation is the single design system package.
