--- packages/animations/CHANGELOG.md
@@ -1,16 +1,0 @@
-# @revealui/animations
-
-## 0.1.1
-
-### Patch Changes
-
-- 0f195e4: SDLC hardening, content overhaul, and cms→admin rename.
-
-  - Promote all CI quality checks from warn-only to hard-fail
-  - Kill banned phrases across 58 files (headless CMS → agentic business runtime)
-  - Rename apps/cms to apps/admin throughout the codebase
-  - Remove proprietary AI providers (Anthropic, OpenAI direct) — keep OpenAI-compatible base
-  - Add Gmail-first email provider to MCP server (Resend deprecated)
-  - Fix CodeQL security alerts (XSS validation, path traversal guard, prototype-safe objects)
-  - Align all coverage thresholds with actual coverage
-  - Add 4 ADRs (dual-database, Fair Source licensing, session-only auth, two-repo model)
