--- packages/animations/package.json
@@ -1,65 +1,0 @@
-{
-  "name": "@revealui/animations",
-  "version": "0.1.1",
-  "description": "Zero-dependency animation library for RevealUI. Spring physics, gestures, variants, and presence animations.",
-  "keywords": [
-    "animation",
-    "spring",
-    "gesture",
-    "motion",
-    "react",
-    "revealui",
-    "typescript"
-  ],
-  "license": "MIT",
-  "devDependencies": {
-    "@testing-library/react": "^16.3.2",
-    "@types/react": "catalog:",
-    "@types/react-dom": "catalog:",
-    "@vitejs/plugin-react": "^6.0.1",
-    "cross-env": "^10.1.0",
-    "dev": "workspace:*",
-    "jsdom": "catalog:",
-    "typescript": "catalog:",
-    "vite": "catalog:",
-    "vite-plugin-dts": "^4.5.4",
-    "vitest": "catalog:"
-  },
-  "engines": {
-    "node": ">=24.13.0"
-  },
-  "exports": {
-    ".": {
-      "types": "./dist/index.d.ts",
-      "import": "./dist/index.js"
-    }
-  },
-  "files": [
-    "dist"
-  ],
-  "main": "./dist/index.js",
-  "module": "./dist/index.js",
-  "peerDependencies": {
-    "react": "^18.0.0 || ^19.0.0",
-    "react-dom": "^18.0.0 || ^19.0.0"
-  },
-  "publishConfig": {
-    "access": "public",
-    "registry": "https://registry.npmjs.org"
-  },
-  "scripts": {
-    "build": "cross-env NODE_ENV=production vite build",
-    "clean": "rm -rf dist",
-    "dev": "vite build --watch --mode development",
-    "lint": "biome check .",
-    "test": "vitest run",
-    "typecheck": "tsc --noEmit"
-  },
-  "type": "module",
-  "types": "./dist/index.d.ts",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/RevealUIStudio/revealui.git",
-    "directory": "packages/animations"
-  }
-}
