--- packages/animations/vite.config.ts
@@ -1,40 +1,0 @@
-import path from 'node:path';
-import react from '@vitejs/plugin-react';
-import { defineConfig } from 'vite';
-import dts from 'vite-plugin-dts';
-
-export default defineConfig({
-  test: {
-    environment: 'jsdom',
-  },
-  plugins: [
-    react(),
-    dts({
-      include: ['src/**/*'],
-      outDir: 'dist',
-      copyDtsFiles: true,
-      insertTypesEntry: true,
-    }),
-  ],
-  build: {
-    lib: {
-      entry: {
-        index: path.resolve(import.meta.dirname, 'src/index.ts'),
-      },
-      formats: ['es'],
-    },
-    rolldownOptions: {
-      external: ['react', 'react-dom', 'react/jsx-runtime', 'react/jsx-dev-runtime'],
-      output: {
-        globals: {
-          react: 'React',
-          'react-dom': 'ReactDOM',
-        },
-        entryFileNames: '[name].js',
-        banner: () => '"use client";',
-      },
-    },
-    sourcemap: true,
-    minify: false,
-  },
-});
